+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# Unreleased
+
 # v0.8.0 - 2026-06-29
 
 * New column: Datetime formatted at timezone (`datetime.TZ.FMT` column)