+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * `-k`/`--key` supports prefix searching with `*`, e.g. `-k addr:*` works
 * Added `--append-columns`
+* Upgrade dependencies
 
 # v0.8.0 - 2026-06-29