+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * `-k`/`--key` supports prefix searching with `*`, e.g. `-k addr:*` works
 * Added `--append-columns`
+* Added several new columns: `latest_` for details on the latest version of that object.
 * Progress report now done always with indicatif
 * Upgrade dependencies & Internal refactoring